moving code around
